arregladas traducciones de hero´s y probando nueva metadata para gacetillas-detalle
