+++ app/[locale]/informacion-turistica/page.tsx
@@ -1,25 +1,13 @@
 import { ComoLlegarSection } from "@/components/03-informacion-turistica/como-llegar";
 import { DelCampoCiudadSection } from "@/components/03-informacion-turistica/del-campo";
 import SliderBsAs from "@/components/03-informacion-turistica/slider-bsas";
+import WrapperInfoTuristica from "@/components/03-informacion-turistica/wrapper-info-turistica";
 import { HeroSection } from "@/components/hero-section";
 import { getTranslations } from "next-intl/server";
 import React from "react";
 
-const InformcionTuristica = async () => {
-  const t = await getTranslations("informacion-turistica");
-
-  return (
-    <section>
-      <HeroSection
-        backgroundImage="/images/informacion-turistica/hero.webp"
-        title1={t("informacion")}
-        title2={t("turistica")}
-      />
-      <DelCampoCiudadSection />
-      <SliderBsAs />
-      <ComoLlegarSection />
-    </section>
-  );
+const InformcionTuristica = () => {
+  return <WrapperInfoTuristica />;
 };
 
 export default InformcionTuristica;
