+++ app/[locale]/nos-acompanian/page.tsx
@@ -1,21 +1,12 @@
 import Anfitriones from "@/components/06-nos-acompanian/anfitriones";
+import WrapperNosAcompanian from "@/components/06-nos-acompanian/wrapper-nos-acompanian";
 import { HeroSection } from "@/components/hero-section";
 import { getTranslations } from "next-intl/server";
 import React from "react";
 
-const NosAcompanian = async () => {
-  const t = await getTranslations("nos-acompanian");
+const NosAcompanian = () => {
 
-  return (
-    <section>
-      <HeroSection
-        title1={t("nos")}
-        title2={t("acompanian")}
-        backgroundImage="/images/nos-acompanian/hero.webp"
-      />
-      <Anfitriones />
-    </section>
-  );
+  return <WrapperNosAcompanian />;
 };
 
 export default NosAcompanian;
