+++ app/[locale]/preguntas-frecuentes/page.tsx
@@ -1,39 +1,11 @@
-"use client";
-import Accordion from "@/components/faq-accordion";
-import { HeroSection } from "@/components/hero-section";
-import { FaqItem } from "@/lib/types";
-import { useTranslations } from "next-intl";
-import React, { useState } from "react";
+import WrapperPreguntasFrecuentes from "@/components/07-preguntas-frecuentes/wrapper-preguntas-frec";
+import React from "react";
 
 const PreguntasFrecuentes = () => {
-  const t = useTranslations("faq");
-
-  const preguntas = t.raw("items") as Array<{
-    pregunta: string;
-    respuesta: string | string[];
-  }>;
-
-  const [faqs, setFaqs] = useState<FaqItem[]>(
-    Array.isArray(preguntas)
-      ? preguntas.map((item) => ({
-          question: item.pregunta,
-          answer: item.respuesta,
-          isOpen: false,
-        }))
-      : [],
-  );
-
   return (
-    <section>
-      <HeroSection
-        title1={t("preguntas")}
-        title2={t("frecuentes")}
-        backgroundImage="/images/preguntas-frecuentes/hero.webp"
-      />
-      <div className="mx-auto max-w-[820px] px-4 py-8">
-        <Accordion items={faqs} addShadow={true} />
-      </div>
-    </section>
+    <>
+      <WrapperPreguntasFrecuentes />
+    </>
   );
 };
