+++ messages/en.json
@@ -326,7 +326,7 @@
     "hero-title": "News",
     "title": "Newsletters",
     "palabras": "Key Words",
-    "día": "Day",
+    "dia": "Day",
     "buscar": "Search",
     "medios": "Media",
     "descargar-gacetilla": "Download gacetilla",
