+++ messages/es.json
@@ -314,7 +314,7 @@
     "hero-title": "Prensa",
     "title": "Gacetillas",
     "palabras": "Palabras claves",
-    "día": "Día",
+    "dia": "Día",
     "buscar": "Buscar",
     "medios": "Medios",
     "descargar-gacetilla": "Descargar gacetilla",
